Extract TestFileHelper into shared test utilities
- Create Tests/XProjectTests/Utilities/TestFileHelper.swift as shared utility
- Remove duplicate TestFileHelper structs from ConfigurationTests and SetupServiceTests
- Consolidate all test utility functions in single location
- Optimize configurationLoaderWithInvalidYAML test to use shared withTemporaryFile utility
- Maintains all existing functionality while eliminating code duplication
- All 66 tests continue to pass with 0 SwiftLint violations

Addresses code review suggestion #3 from PR feedback
