+++ Tests/XProjectTests/Services/SetupServiceTests.swift
@@ -9,22 +9,6 @@ import Testing
 
 // MARK: - Test Helpers
 
-private struct TestFileHelper {
-    @discardableResult
-    static func ensureDummyProject(at path: URL, name: String = "DummyProject") -> URL {
-        let projectURL = path.appendingPathComponent("\(name).xcodeproj")
-        if !FileManager.default.fileExists(atPath: projectURL.path) {
-            let dummyContent = """
-                // Dummy project file for testing
-                // This file exists solely to provide a valid project path for tests
-                // that need to reference an existing file during test execution
-                """
-            try? dummyContent.write(to: projectURL, atomically: true, encoding: .utf8)
-        }
-        return projectURL
-    }
-}
-
 private struct ConfigurationTestHelper {
     static func createTestConfigurationService() -> ConfigurationService {
         let configPath = Bundle.module.path(forResource: "test-config", ofType: "yml", inDirectory: "Support")!
